let rating counts be 32 bit integers

+++ src/lib/incident/incidents-service.ts
@@ -188,8 +188,8 @@ export async function getIncidentRating(
   const res = await getConnection().query(
     `
     SELECT
-      count(*) FILTER ( WHERE rating = 1 ) "positiveCount",
-      count(*) FILTER ( WHERE rating = -1 ) "negativeCount",
+      (count(*) FILTER ( WHERE rating = 1 ))::INT "positiveCount",
+      (count(*) FILTER ( WHERE rating = -1 ))::INT "negativeCount",
       max(created_at) "lastRatedAt"
     FROM incident_ratings
     WHERE incident_id = $1
